msg, nth, nua, sip, soa, sresolv, su, tport: signed/unsigned fixes, fixes in checking error returns.
darcs-hash:20060926112825-65a35-d94b2830f4e2cc4d7af21308f69d4d94d068b3e6.gz

+++ libsofia-sip-ua/su/sofia-sip/su_alloc.h
@@ -81,9 +81,9 @@ SU_DLL void su_home_preload(su_home_t *h, isize_t n, isize_t size);
 SU_DLL su_home_t *su_home_auto(void *area, isize_t size);
 
 #define SU_HOME_AUTO_SIZE(n)				\
-  (((n) + ((sizeof(su_home_t) + 7) & ~8) +		\
+  (((n) + ((sizeof(su_home_t) + 7) & (size_t)~8) +	\
     ((3 * sizeof (void *) + 4 * sizeof(unsigned) +	\
-      7 * (sizeof (long) + sizeof(void *)) + 7) & ~8)) 	\
+      7 * (sizeof (long) + sizeof(void *)) + 7) & (size_t)~8)) 	\
     / sizeof(su_home_t))
 
 SU_DLL int su_home_move(su_home_t *dst, su_home_t *src);
